+++ Dockerfiles/build.sh
@@ -1,4 +1,4 @@
-tag=0.10.9
+tag=0.11.0
 docker build -f Dockerfile.$tag -t rnakato/churros:$tag . #--no-cache
 #docker save -o churros-$tag.tar rnakato/churros:$tag
 #singularity build -F churros.$tag.sif docker-archive://churros-$tag.tar